musicplayer stability improvement
