+++ cogs/music/__init__.py
@@ -1,5 +1,5 @@
-from .music import Voice
+from .commands import MusicCommands
 
 async def setup(client) -> None:
     """Adds the cog to the client"""
-    await client.add_cog(Voice(client))
+    await client.add_cog(MusicCommands(client))
