+++ cogs/music/song.py
